feat: Update migrations to use integer types
price and amount  for Accurate, fast, cross-currency safe

+++ database/migrations/2025_11_13_120027_create_order_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained()->cascadeOnDelete();
 
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
-            $table->decimal('amount_per_item', 10, 2);
+            $table->integer('amount_per_item');
             $table->unsignedInteger('quantity');
             
             $table->timestamps();
